Рефакторит видео-плеер.
