+++ src/hocs/with-video-player/with-video-player.jsx
@@ -1,5 +1,6 @@
 import React, {PureComponent} from 'react';
-import {PlayerWithVideo, VideoPlayerMode} from '../../components/player/player';
+import {PlayerWithVideo} from '../../components/player/player';
+import {VideoPlayerMode, videoOptions} from "../with-video/with-video";
 
 
 export const VideoPlayerStatus = {
@@ -22,20 +23,6 @@ export const withVideoPlayer = (Component) => {
       this._playerMode = null;
     }
 
-    getPlayerOptions(playerMode) {
-      switch (playerMode) {
-        case VideoPlayerMode.PREVIEW:
-          return {
-            isAutoPlay: false,
-            isSound: false,
-            videoHeight: 175,
-          };
-        case VideoPlayerMode.FULL_SCREEN:
-        default:
-          return null;
-      }
-    }
-
     setVideoPlayerStatus(newPlayerStatus) {
       this.setState({
         playerStatus: newPlayerStatus,
@@ -51,7 +38,7 @@ export const withVideoPlayer = (Component) => {
     getPlayingValue() {
       switch (this.state.playerStatus) {
         case VideoPlayerStatus.ON_AUTOPLAY:
-          return this.getPlayerOptions(this._playerMode).isAutoPlay;
+          return videoOptions[this._playerMode].isAutoPlay;
         case VideoPlayerStatus.ON_PLAY:
           return true;
         case VideoPlayerStatus.ON_PAUSE:
@@ -65,17 +52,12 @@ export const withVideoPlayer = (Component) => {
     renderPlayer(src, posterUrl, playerMode) {
       this._playerMode = playerMode;
 
-      const options = this.getPlayerOptions(playerMode);
-      const isPlaying = this.getPlayingValue();
-
       return (
         <PlayerWithVideo
           src={src}
           posterUrl={posterUrl}
-          videoHeight={options.videoHeight}
           playerMode={playerMode}
-          isPlaying={isPlaying}
-          isSound={options.isSound}
+          isPlaying={this.getPlayingValue()}
         />
       );
     }
